Added function to generate file names
The plan is to incorporate the CBOR file generation for each sensor batch reading.  But, I have not been able to get console log sensor readings using the toggle button.  Does the sensor.start() function call run the if Accelerometer block or would that need to be in an Accelerometer function?  In fact, only heart rate outputs are going to the console, none of the batch readings.